+++ volcengine-java-sdk-billing/src/main/java/com/volcengine/billing/model/DeleteFinancialRelationRequest.java
@@ -37,7 +37,7 @@ public class DeleteFinancialRelationRequest {
   private String relationID = null;
 
   @SerializedName("SubAccountID")
-  private Integer subAccountID = null;
+  private Long subAccountID = null;
 
   public DeleteFinancialRelationRequest relation(Integer relation) {
     this.relation = relation;
@@ -77,7 +77,7 @@ public class DeleteFinancialRelationRequest {
     this.relationID = relationID;
   }
 
-  public DeleteFinancialRelationRequest subAccountID(Integer subAccountID) {
+  public DeleteFinancialRelationRequest subAccountID(Long subAccountID) {
     this.subAccountID = subAccountID;
     return this;
   }
@@ -88,11 +88,11 @@ public class DeleteFinancialRelationRequest {
   **/
   @NotNull
   @Schema(required = true, description = "")
-  public Integer getSubAccountID() {
+  public Long getSubAccountID() {
     return subAccountID;
   }
 
-  public void setSubAccountID(Integer subAccountID) {
+  public void setSubAccountID(Long subAccountID) {
     this.subAccountID = subAccountID;
   }
